Version 1.0.0-alpha
We are changing the major classifier's concept.
It is now based on a space transformation, that is achieved by measuring the
distance between a sample and a representant of each class, so the data
is mapped from N-d to C-d, where C is the number of classes.

Then, the hierarchy of the classifier is built primarily on dividing the new
space, using the quadtree algorithm, considering the number of samples of each
class as the homogeneity criteria. Besides, there is a modification to merge
the empty quadrants with non-empty ones.

The classification step consists of mapping and then finding in which
quadrant the new sample lies in.

Log:
 - created the QuadTreeNode class;
 - created the SpaceRegion class;
 - created the new HierarchicalClassifier class;
 - implemented the fit method of the new classifier;
 - implemented the merge method;
 - implemented the predict method;

Signed-off-by: Paulo Hack <ph_hack@hotmail.com>